bioconda_utils/bioconda_utils-requirements.txt
# Some of Bioconda's supported versions are described
# in ../docs/source/user/versions.rst
# so please keep that documentation up to date as they change

# basics
python>=3.7
conda=23.3.*
boa=0.15.*
conda-build=3.24.*
conda-verify=3.1.*
argh=0.26.*          # CLI
colorlog=4.8.*       # Logging
tqdm>=4.26           # Progress monitor
yaspin =2
ruamel_yaml=0.15.*   # Recipe YAML parsing
pyaml=17.12.*        # Faster YAML parser (deprecate?)
networkx=2.*
pandas=1.4.*
numpy=1.19.*         # Avoid breaking pandas on OSX
libblas=*=*openblas  # Avoid large mkl package (pulled in by pandas)
boltons=23.*
jsonschema=3.2.*     # JSON schema verification
pyopenssl>=22.1      # Stay compatible with cryptography

# pinnings
conda-forge-pinning=2023.05.06.13.08.41

# tools
anaconda-client=1.6.*  # anaconda_upload
involucro=1.1.*        # mulled test and container build
skopeo=1.11.*          # docker upload
git=2.*                # well - git

# hosters - special regex not supported by RE
regex=2022.7.9

# asyncio
aiohttp=3.8.*      # HTTP lib
aiohttp-jinja2     # jinja2 renderer for aiohttp.web
aiohttp-session
aiohttp-security
aiofiles=0.8.*     # async open
aioftp=0.12.*      # FTP lib
backoff=1.6.*      # retrying of failed API calls
cachetools=3.0.*   # request caching (NEEDED?)

# client API wrappers
gitpython>=3.0.8,3.0.*    # githandler
                          # needs >=3.0.8 due to https://github.com/conda-forge/staged-recipes/issues/10874
gidgethub=3.0.*           # githubhandler
pyjwt>=2.4.0              # githubhandler (JWT signing), needs >=2.4.0, CVE-2022-29217

# unknown
beautifulsoup4=4.8.*
galaxy-lib>=18.9.1
jinja2>=2.10.1,<3
markupsafe<2.1           # markupsafe 2.1 breaks jinja2

# docs
sphinx>=4.1
celery
sphinx-autodoc-typehints
alabaster=0.7.*
docutils
markdown
graphviz

# The bioconductor skeleton needs this
requests=2.22.*

# merge handling
pygithub

# caching
diskcache =5.*
appdirs =1.*

# build failure output
tabulate =0.9